refactor: centralize item sorting logic
