+++ Intersect.Client.Core/Interface/Game/Bank/BankWindow.cs
@@ -238,7 +238,7 @@ public partial class BankWindow : Window
         {
             for (var i = 0; i < max - pass - 1; i++)
             {
-                if (CompareSlots(bank, i, i + 1) > 0)
+                if (ItemListHelper.CompareItems(bank[i], bank[i + 1], _criterion, _sortAscending) > 0)
                 {
                     PacketSender.SendMoveBankItems(i, i + 1);
                     (bank[i], bank[i + 1]) = (bank[i + 1], bank[i]);
@@ -247,163 +247,6 @@ public partial class BankWindow : Window
         }
 
         ApplyFilters();
-    }
-
-    private int CompareSlots(Item?[] bank, int a, int b)
-    {
-        var left = (a >= 0 && a < bank.Length) ? bank[a] : null;
-        var right = (b >= 0 && b < bank.Length) ? bank[b] : null;
-
-        var dx = left?.Descriptor;
-        var dy = right?.Descriptor;
-
-        var leftEmpty = dx == null;
-        var rightEmpty = dy == null;
-
-        if (leftEmpty && rightEmpty) return 0;
-        if (leftEmpty) return 1;
-        if (rightEmpty) return -1;
-
-        return _criterion switch
-        {
-            SortCriterion.Name =>
-                CompareByName(dx!, dy!, _sortAscending),
-            SortCriterion.Quantity =>
-                CompareQuantity(left!, right!),
-            SortCriterion.Price =>
-                ComparePrice(dx!, dy!),
-            _ => CompareDescriptors(dx!, dy!, _sortAscending)
-        };
-    }
-
-    private int CompareByName(ItemDescriptor dx, ItemDescriptor dy, bool ascending)
-    {
-        var cmp = StringComparer.OrdinalIgnoreCase.Compare(SafeName(dx), SafeName(dy));
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(dx, dy, ascending);
-    }
-
-    private int CompareQuantity(Item left, Item right)
-    {
-        var qx = left.Quantity;
-        var qy = right.Quantity;
-
-        var cmp = qx.CompareTo(qy);
-        if (cmp != 0)
-        {
-            return _sortAscending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(left.Descriptor!, right.Descriptor!, _sortAscending);
-    }
-
-    private int ComparePrice(ItemDescriptor dx, ItemDescriptor dy)
-    {
-        var px = dx.Price;
-        var py = dy.Price;
-
-        var cmp = px.CompareTo(py);
-        if (cmp != 0)
-        {
-            return _sortAscending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(dx, dy, _sortAscending);
-    }
-
-    private int CompareDescriptors(ItemDescriptor x, ItemDescriptor y, bool ascending)
-    {
-        var cmp = TypePriority(x.ItemType).CompareTo(TypePriority(y.ItemType));
-        if (cmp != 0)
-        {
-            return cmp;
-        }
-
-        cmp = SubtypeIndex(x).CompareTo(SubtypeIndex(y));
-        if (cmp != 0)
-        {
-            return cmp;
-        }
-
-        cmp = StringComparer.OrdinalIgnoreCase.Compare(SafeName(x), SafeName(y));
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        cmp = x.Price.CompareTo(y.Price);
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        return x.Id.CompareTo(y.Id);
-    }
-
-    private int TypePriority(ItemType t)
-    {
-        return t switch
-        {
-            ItemType.Currency => 0,
-            ItemType.Equipment => 10,
-            ItemType.Bag => 20,
-            ItemType.Event => 30,
-            ItemType.Spell => 40,
-            ItemType.Consumable => 50,
-            ItemType.Resource => 60,
-            _ => 100 + (int)t
-        };
-    }
-
-    private int SubtypeIndex(ItemDescriptor d)
-    {
-        try
-        {
-            var dict = Options.Instance.Items.ItemSubtypes;
-            if (dict == null)
-            {
-                return int.MaxValue;
-            }
-
-            var targetType = Convert.ToInt32(d.ItemType);
-
-            foreach (var kv in dict)
-            {
-                if (Convert.ToInt32(kv.Key) == targetType)
-                {
-                    var list = kv.Value;
-                    if (list == null)
-                    {
-                        break;
-                    }
-
-                    var st = d.Subtype ?? string.Empty;
-                    for (var i = 0; i < list.Count; i++)
-                    {
-                        if (string.Equals(list[i], st, StringComparison.OrdinalIgnoreCase))
-                        {
-                            return i;
-                        }
-                    }
-
-                    break;
-                }
-            }
-        }
-        catch
-        {
-        }
-
-        return int.MaxValue;
-    }
-
-    private static string SafeName(ItemDescriptor d)
-    {
-        return d.Name ?? string.Empty;
     }
     private void UpdateSubtypeOptions()
     {
