+++ Intersect.Client.Core/Interface/Game/Inventory/InventoryWindow.cs
@@ -341,7 +341,7 @@ public partial class InventoryWindow : Window
         {
             for (var i = 0; i < max - pass - 1; i++)
             {
-                if (CompareSlots(inventory, i, i + 1) > 0)
+                if (ItemListHelper.CompareItems(inventory[i], inventory[i + 1], _criterion, _sortAscending) > 0)
                 {
                     Globals.Me.SwapItems(i, i + 1);
                 }
@@ -349,178 +349,6 @@ public partial class InventoryWindow : Window
         }
 
         ApplyFilters();
-    }
-
-    /// <summary>
-    /// Compara los contenidos de los slots 'a' y 'b'.
-    /// Devuelve &gt; 0 si A debe ir después de B, &lt; 0 si A va antes, 0 si empate.
-    /// </summary>
-    private int CompareSlots(IItem?[] inventory, int a, int b)
-    {
-        var left = (a >= 0 && a < inventory.Length) ? inventory[a] : null;
-        var right = (b >= 0 && b < inventory.Length) ? inventory[b] : null;
-
-        var dx = left?.Descriptor;
-        var dy = right?.Descriptor;
-
-        var leftEmpty = dx == null;
-        var rightEmpty = dy == null;
-
-        // Vacíos siempre al final
-        if (leftEmpty && rightEmpty) return 0;
-        if (leftEmpty) return 1;
-        if (rightEmpty) return -1;
-
-        return _criterion switch
-        {
-            SortCriterion.Name =>
-                CompareByName(dx!, dy!, _sortAscending),
-            SortCriterion.Quantity =>
-                CompareQuantity(left!, right!),
-            SortCriterion.Price =>
-                ComparePrice(dx!, dy!),
-            _ => CompareDescriptors(dx!, dy!, _sortAscending)
-        };
-    }
-
-    private int CompareByName(ItemDescriptor dx, ItemDescriptor dy, bool ascending)
-    {
-        var cmp = StringComparer.OrdinalIgnoreCase.Compare(SafeName(dx), SafeName(dy));
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(dx, dy, ascending);
-    }
-
-    private int CompareQuantity(IItem left, IItem right)
-    {
-        var qx = left.Quantity;
-        var qy = right.Quantity;
-
-        var cmp = qx.CompareTo(qy);
-        if (cmp != 0)
-        {
-            return _sortAscending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(left.Descriptor!, right.Descriptor!, _sortAscending);
-    }
-
-    private int ComparePrice(ItemDescriptor dx, ItemDescriptor dy)
-    {
-        var px = dx.Price;
-        var py = dy.Price;
-
-        var cmp = px.CompareTo(py);
-        if (cmp != 0)
-        {
-            return _sortAscending ? cmp : -cmp;
-        }
-
-        return CompareDescriptors(dx, dy, _sortAscending);
-    }
-
-    /// <summary>
-    /// Orden "TypeThenName": Tipo → Subtipo → Nombre → Precio → Id.
-    /// </summary>
-    private int CompareDescriptors(ItemDescriptor x, ItemDescriptor y, bool ascending)
-    {
-        var cmp = TypePriority(x.ItemType).CompareTo(TypePriority(y.ItemType));
-        if (cmp != 0)
-        {
-            return cmp;
-        }
-
-        cmp = SubtypeIndex(x).CompareTo(SubtypeIndex(y));
-        if (cmp != 0)
-        {
-            return cmp;
-        }
-
-        cmp = StringComparer.OrdinalIgnoreCase.Compare(SafeName(x), SafeName(y));
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        cmp = x.Price.CompareTo(y.Price);
-        if (cmp != 0)
-        {
-            return ascending ? cmp : -cmp;
-        }
-
-        return x.Id.CompareTo(y.Id);
-    }
-
-    /// <summary>
-    /// Prioridad del tipo de ítem para agrupar de forma estable.
-    /// </summary>
-    private int TypePriority(ItemType t)
-    {
-        return t switch
-        {
-            ItemType.Currency => 0,
-            ItemType.Equipment => 10,
-            ItemType.Bag => 20,
-            ItemType.Event => 30,
-            ItemType.Spell => 40,
-            ItemType.Consumable => 50,
-            ItemType.Resource => 60,
-            _ => 100 + (int)t
-        };
-    }
-
-    /// <summary>
-    /// Obtiene la posición del subtipo en la configuración.
-    /// </summary>
-    private int SubtypeIndex(ItemDescriptor d)
-    {
-        try
-        {
-            var dict = Options.Instance.Items.ItemSubtypes;
-            if (dict == null)
-            {
-                return int.MaxValue;
-            }
-
-            var targetType = Convert.ToInt32(d.ItemType);
-
-            foreach (var kv in dict)
-            {
-                if (Convert.ToInt32(kv.Key) == targetType)
-                {
-                    var list = kv.Value;
-                    if (list == null)
-                    {
-                        break;
-                    }
-
-                    var st = d.Subtype ?? string.Empty;
-                    for (var i = 0; i < list.Count; i++)
-                    {
-                        if (string.Equals(list[i], st, StringComparison.OrdinalIgnoreCase))
-                        {
-                            return i;
-                        }
-                    }
-
-                    break;
-                }
-            }
-        }
-        catch
-        {
-            // Ignorar errores y mandar al final.
-        }
-
-        return int.MaxValue;
-    }
-
-    private static string SafeName(ItemDescriptor d)
-    {
-        return d.Name ?? string.Empty;
     }
 
     protected override void EnsureInitialized()
